Updated RoomFile.cs; Commented Out Empty Methods
I'm Still trying to figure out why I'm getting a read error. Another issue was that the empty methods in PlayerFile.cs and PlayerCreation.cs were what caused the program not to build when Sprint 2 was submitted. They will need to stay commented out until we put something in them.